% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.exact.R
\name{confint.RMA.Exact}
\alias{confint.RMA.Exact}
\title{Compute a confidence interval for the population mean from an RMA.Exact object}
\usage{
\method{confint}{RMA.Exact}(rma0, levels = 0.05, ...)
}
\arguments{
\item{rma0}{an object of class RMA.Exact}

\item{levels}{the significance levels at which the compute the confidence intervals}

\item{...}{(currently for internal use)}
}
\value{
a matrix with a row corresponding to each weight parameter c0 stored in rma0, and columns containing the upper and lower interval endpoints, the population variance values at whcih those
endpoints were obtained, etc.
}
\description{
A warning will be issued if the reported endpoints of the confidence interval are near the bounding box.
}
\examples{
see ?RMA.Exact
}
