% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.exact.R
\name{plot.RMA.Exact}
\alias{plot.RMA.Exact}
\title{Plot a confidence region given by an RMA.Exact object}
\usage{
\method{plot}{RMA.Exact}(rma0, levels = c(0.01, 0.02, 0.05, 0.1, 0.15, 0.2),
  mfrow = c(1, dim(rma0)[3]), ...)
}
\arguments{
\item{rma0}{an object of class RMA.Exact}

\item{levels}{the significance levels to plot}

\item{mfrow}{the dimensions of the array of plotting windows for use when rma0 contains regions for multiple weight parameters; defaults to a single row with as many as columns as regions
contained in rma0}

\item{...}{additional parameters passed down to graphics::contour (passed from there down to plot.window)}
}
\value{
undefined
}
\description{
Plot a confidence region given by an RMA.Exact object
}
\examples{
see ?RMA.Exact
}
