% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.exact.R
\name{+.RMA.Exact}
\alias{+.RMA.Exact}
\title{Combine confidence regions from two RMA.Exact objects}
\usage{
\method{+}{RMA.Exact}(rma1, rma2)
}
\arguments{
\item{rma1}{the first RMA.Exact object}

\item{rma2}{the second RMA.Exact object}
}
\value{
an object of class RMA.Exact
}
\description{
Combine confidence regions from two RMA.Exact objects
}
