% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.exact.R
\name{rma.exact.fast}
\alias{rma.exact.fast}
\title{Compute a confidence interval for the grand mean at a user-specified level.}
\usage{
rma.exact.fast(yi, vi, c0 = 1, level = 0.05, plot = TRUE,
  tau2.bounds = NULL, resolution = 100, Z = NULL, B = 3000,
  tau2.alpha = 0.995)
}
\arguments{
\item{yi}{a vector containing the primary study measurements}

\item{vi}{a vector of the same length as yi containing the variances of the of the primary study measurements contained in yi}

\item{c0}{a vector containing the mixing parameters for the test statistics; defaults to 1}

\item{level}{the level of the confidence interval; defaults to .05}

\item{plot}{indicator whether to plot the contour of the confidence region; defaults to TRUE}

\item{tau2.bounds}{upper and lower bounds for the range of population variance values for constructing the confidence region; if NULL, value will be calculated from tau2.alpha}

\item{resolution}{resolution of the population variance values for constructing the confidence region; defaults to 1e2}

\item{Z}{a matrix of length(yi) rows with each row consisting of standard normal samples to be used in the monte carlo estimation of the null distribution of the test statistic; if NULL, B values
will be sampled per row}

\item{B}{the number of monte carlo replicates per primary study observation to be used; defaults to 300}

\item{tau2.alpha}{the level of the exact CI with which to bounds on population variance when constructing the confidence region}
}
\value{
a matrix with length(c0) rows and each row containing the lower and upper endpoints of the confidence interval for the given mixing parameter
}
\description{
Compute a confidence interval for the grand mean at a user-specified level.
}
\examples{

set.seed(1)
K <- 4
c0 <- c(.5,1)
yi=rnorm(K)*sqrt(vi+tau2)+mu0
vi <- (seq(1, 5, length=K))^2
Z <- matrix(rnorm(K*5e3),nrow=K)
rma.exact.fast(yi,vi,resolution=5e2)
}
\seealso{
\code{\link{rma.exact}} for computing entire confidence regions
}
